Messed with function call code
Now mMapFun stores a function name and a pair of function pointer and argument number associated with it.

Now every function gets page and player as an argument, plus a new argument -- std::vector<std::string>.
I'm not sure what direction you wanted to take with function arguments, so you might want change this vector of strings to something else.